Am incheiat de implementat ierarhia de produse, am implementat ierarhia de magazine, si clasa mall

+++ Proiect_Mall/Proiect_Mall/product.cpp
@@ -21,15 +21,10 @@ bool Product::Sell() {
 	}
 }
 
-const int Product::GetInitialSellPrice() const {
+const float Product::GetInitialSellPrice() const {
 	return InitialSellPrice;
 }
-const int Product::GetBuyPrice() const {
+const float Product::GetBuyPrice() const {
 	return BuyPrice;
 }
-
-
-
-
-
